More central architecture
+ Enemy health bars
+ Player Controller
+ Store Mouse position in PlayerStateMachine
+ Take damage on Enemy editor script

+++ LRJ2022Project/Assets/Scripts/Enemies/EnemyData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu]
 public class EnemyData : ScriptableObject
 {
-    public int MaxHealth;
+    public float MaxHealth;
     public int Damage;
     public float MoveSpeed;
     public float MoveHeight;
